Fix sentence loading in run_tagger.
Previously, multiple empty lines would be considered as one sentence
ending and therefore the number of lines on output could be smaller than
on input.

Now the number of lines on output is equal the number lines on input,
plus one if the last input sentence does not end with empty line.