Remove english_elementary_features.
The generic_elementary_features is a strict superset and English model
was never released publically.

Nevertheless, we keep the value of tagger_ids::ENGLISH3 (which is now
removed) unassigned, so that if some old model is around, it cannot be
loaded (instead of loading and crashing, or loading and producing wrong
results).