Add raw_form_len API call.
The call returns raw_form length of a given form, analogously
to raw_lemma_len. Nevetheless, it is used only because of
external_morpho, which needs to known the structure of the input form
when tagging to use only the real form without the analyses. Otherwise
the identity of a form would depend on the analyses and their order.