Fix tokenization of long sentences.
Fixes #6.

When a sentence is too long and is forcibly ended using
emergency_sentence_split, one character (right after the last token)
was skipped (in XML output, it was outside the <sentence>s).

The problem was caused by Ragel advancing the variable current
during fbreak. Using an additional fhold fixes the problem.